Move Activity provider to  folder
