+++ src/App.js
@@ -9,7 +9,7 @@ import OnboardingLoader from './components/OnboardingLoader'
 import EmailNotificationsLoader from './components/EmailNotificationsLoader'
 import RequestPanel from './components/RequestPanel/RequestPanel'
 import Routes from './Routes'
-import { ActivityProvider } from './components/Activity/ActivityProvider'
+import { ActivityProvider } from './providers/ActivityProvider'
 import { CourtClockProvider } from './providers/CourtClock'
 import { CourtConfigProvider } from './providers/CourtConfig'
 import { RequestQueueProvider } from './providers/RequestQueue'
